break-into-modules: rename commons.commons to commons.essentials
