+++ apigateway/src/main/java/nva/commons/apigateway/ApiGatewayHandler.java
@@ -2,7 +2,7 @@ package nva.commons.apigateway;
 
 
 import static nva.commons.apigateway.ContentTypes.APPLICATION_JSON;
-import static nva.commons.commons.JsonUtils.objectMapper;
+import static nva.commons.essentials.JsonUtils.objectMapper;
 import com.amazonaws.services.lambda.runtime.Context;
 import java.io.BufferedWriter;
 import java.io.IOException;
@@ -19,8 +19,8 @@ import nva.commons.apigateway.exceptions.ApiGatewayException;
 import nva.commons.apigateway.exceptions.ApiGatewayUncheckedException;
 import nva.commons.apigateway.exceptions.GatewayResponseSerializingException;
 import nva.commons.apigateway.exceptions.LoggerNotSetException;
-import nva.commons.commons.Environment;
-import nva.commons.commons.JsonUtils;
+import nva.commons.essentials.Environment;
+import nva.commons.essentials.JsonUtils;
 
 
 import org.slf4j.Logger;
