+++ apigateway/src/test/java/nva/commons/apigateway/ApiGatewayHandlerTest.java
@@ -3,7 +3,7 @@ package nva.commons.apigateway;
 import static nva.commons.apigateway.ApiGatewayHandler.APPLICATION_PROBLEM_JSON;
 import static nva.commons.apigateway.ApiGatewayHandler.CONTENT_TYPE;
 import static nva.commons.apigateway.ApiGatewayHandler.REQUEST_ID;
-import static nva.commons.commons.JsonUtils.objectMapper;
+import static nva.commons.essentials.JsonUtils.objectMapper;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.core.Is.is;
@@ -35,7 +35,7 @@ import nva.commons.apigateway.exceptions.InvalidOrMissingTypeException;
 import nva.commons.apigateway.exceptions.TestException;
 import nva.commons.apigateway.testutils.Handler;
 import nva.commons.apigateway.testutils.RequestBody;
-import nva.commons.commons.Environment;
+import nva.commons.essentials.Environment;
 
 import nva.commons.ioutils.IoUtils;
 import nva.commons.logutils.LogUtils;
