+++ apigateway/src/test/java/nva/commons/apigateway/VoidTest.java
@@ -13,9 +13,9 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Path;
 import nva.commons.apigateway.exceptions.ApiGatewayException;
-import nva.commons.commons.Environment;
+import nva.commons.essentials.Environment;
 
-import nva.commons.commons.JsonUtils;
+import nva.commons.essentials.JsonUtils;
 import nva.commons.ioutils.IoUtils;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.BeforeEach;
