+++ essentials/src/test/java/nva/commons/essentials/StringUtilsTest.java
@@ -1,4 +1,4 @@
-package nva.commons.commons;
+package nva.commons.essentials;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
