+++ lambdaauthorizer/src/main/java/no/unit/commons/apigateway/authentication/AuthorizerResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import java.io.ByteArrayOutputStream;
 import java.nio.charset.StandardCharsets;
-import nva.commons.commons.JsonUtils;
+import nva.commons.essentials.JsonUtils;
 
 public class AuthorizerResponse {
